PR #5376: Clang-CL Windows build-fixes (Contributed by TBBle)
#rb none

[CL 4677246 by Ben Marsh in Dev-Build branch]

+++ Engine/Source/Developer/Windows/ShaderFormatD3D/Private/D3D11ShaderCompiler.cpp
@@ -926,7 +926,7 @@ static bool CompileAndProcessD3DShader(FString& PreprocessedShaderSource, const 
 
 			// append data that is generate from the shader code and assist the usage, mostly needed for DX12 
 			{
-				FShaderCodePackedResourceCounts PackedResourceCounts = { bGlobalUniformBufferUsed, NumSamplers, NumSRVs, NumCBs, NumUAVs };
+				FShaderCodePackedResourceCounts PackedResourceCounts = { bGlobalUniformBufferUsed, static_cast<uint8>(NumSamplers), static_cast<uint8>(NumSRVs), static_cast<uint8>(NumCBs), static_cast<uint8>(NumUAVs) };
 
 				Output.ShaderCode.AddOptionalData(PackedResourceCounts);
 				Output.ShaderCode.AddOptionalData('u', UniformBufferNameBytes.GetData(), UniformBufferNameBytes.Num());
